Update main.py
In this commit we:
Imported pdfplumber, created 3 functions to open, read and store the text from the imported pdf. I also fixed the "Read PDF" button and it is functional now. The result of the current code is able to extract the text from a pdf and store it under the pdf_text variable.